+++ roms_tools/datasets/download.py
@@ -83,6 +83,7 @@ pup_test_data = pooch.create(
         "coarsened_UNIFIED_bgc_dataset.nc": "sha256:269d5bcd8e6e64d3400362ae0a65afe049810ce06c536ccf31cca7c00f321bc1",
         "WOA_2018_quarterDeg_coarsened.nc": "sha256:673ce3c3a98bb386ccd899dbc23eeedf7d9a665b68ea52c96fd69829a4b929a7",
         "mbl_co2_bgc_dataset.nc": "sha256:797a9ef48f3c83a6920e44c0b441feb00fb35553db09dea9ed4ff36dcd68d968",
+        "coarsened_OceanSODA_dataset.nc": "sha256:b4a284303c9c1a8904a6ea3fa338ff05ac042c0b024fb4e1ad4bbb1d9fedff38",
         "grid_created_with_matlab.nc": "fd537ef8159fabb18e38495ec8d44e2fa1b7fb615fcb1417dd4c0e1bb5f4e41d",
         "etopo5_coarsened_and_shifted.nc": "9a5cb4b38c779d22ddb0ad069b298b9722db34ca85a89273eccca691e89e6f96",
         "srtm15_coarsened.nc": "48bc8f4beecfdca9c192b13f4cbeef1455f49d8261a82563aaec5757e100dff9",